# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RASHED — AI Water Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /household:
    get:
      operationId: getHousehold
      tags: [household]
      summary: Get household profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
        "404":
          description: No household yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertHousehold
      tags: [household]
      summary: Create or update household
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HouseholdInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
  /household/members:
    get:
      operationId: listHouseholdMembers
      tags: [household]
      summary: List household members
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HouseholdMember"
    post:
      operationId: addHouseholdMember
      tags: [household]
      summary: Add a household member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HouseholdMemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseholdMember"
  /household/members/{id}:
    delete:
      operationId: removeHouseholdMember
      tags: [household]
      summary: Remove a household member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── METERS ───
  /meters:
    get:
      operationId: listMeters
      tags: [meters]
      summary: List meters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meter"
    post:
      operationId: createMeter
      tags: [meters]
      summary: Create a meter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meter"
  /meters/{id}:
    get:
      operationId: getMeter
      tags: [meters]
      summary: Get a meter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meter"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMeter
      tags: [meters]
      summary: Update a meter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeterUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meter"
    delete:
      operationId: deleteMeter
      tags: [meters]
      summary: Delete a meter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /analysis/latest:
    get:
      operationId: getLatestAnalysis
      tags: [analysis]
      summary: Get the latest consumption analysis
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumptionAnalysis"
        "404":
          description: No analysis yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Household:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        memberCount:
          type: number
        propertyType:
          type: string
          enum: [villa, apartment, townhouse, other]
        bathroomCount:
          type: number
        hasGarden:
          type: boolean
        hasPool:
          type: boolean
        propertySizeM2:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - memberCount
        - propertyType
        - bathroomCount
        - hasGarden
        - hasPool
        - createdAt
    HouseholdInput:
      type: object
      properties:
        name:
          type: string
        memberCount:
          type: number
          minimum: 1
        propertyType:
          type: string
          enum: [villa, apartment, townhouse, other]
        bathroomCount:
          type: number
          minimum: 1
        hasGarden:
          type: boolean
        hasPool:
          type: boolean
        propertySizeM2:
          type: ["number", "null"]
      required:
        - name
        - memberCount
        - propertyType
        - bathroomCount
        - hasGarden
        - hasPool
    HouseholdMember:
      type: object
      properties:
        id:
          type: number
        householdId:
          type: number
        name:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - householdId
        - name
        - role
        - createdAt
    HouseholdMemberInput:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
      required:
        - name
        - role

    # ─── METERS ───
    Meter:
      type: object
      properties:
        id:
          type: number
        householdId:
          type: number
        meterId:
          type: string
        meterType:
          type: string
          enum: [main, garden, pool, other]
        label:
          type: ["string", "null"]
        isActive:
          type: boolean
        lastReadingValue:
          type: ["number", "null"]
        lastReadingDate:
          type: ["string", "null"]
        dataClassification:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - householdId
        - meterId
        - meterType
        - isActive
        - dataClassification
        - createdAt
    MeterInput:
      type: object
      properties:
        meterId:
          type: string
        meterType:
          type: string
          enum: [main, garden, pool, other]
        label:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - meterId
        - meterType
    MeterUpdate:
      type: object
      properties:
        label:
          type: ["string", "null"]
        isActive:
          type: boolean
...
    ConsumptionAnalysis:
      type: object
      properties:
        id:
          type: number
        billId:
          type: number
        currentConsumptionM3:
          type: ["number", "null"]
        previousConsumptionM3:
          type: ["number", "null"]
        changePercent:
          type: ["number", "null"]
        anomalyDetected:
          type: boolean
        anomalyExplanation:
          type: ["string", "null"]
        leakRiskLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        leakRiskProbability:
          type: ["number", "null"]
        leakRiskReason:
          type: ["string", "null"]
        whyConsumptionIncreased:
          type: array
          items:
            $ref: "#/components/schemas/ConsumptionCause"
        dataClassification:
          type: string
        tariffConfigVersion:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - billId
        - anomalyDetected
        - whyConsumptionIncreased
        - dataClassification
        - createdAt
    ConsumptionCause:
      type: object
      properties:
        cause:
          type: string
        confidence:
          type: string
          enum: [low, medium, high]
        explanation:
          type: string
      required:
        - cause
        - confidence
        - explanation

    # ─── CONSUMPTION ───
...