# lib/api-zod/src/generated/types/meter.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RASHED — AI Water Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { MeterMeterType } from './meterMeterType';

export interface Meter {
  id: number;
  householdId: number;
  meterId: string;
  meterType: MeterMeterType;
  /** @nullable */
  label?: string | null;
  isActive: boolean;
  /** @nullable */
  lastReadingValue?: number | null;
  /** @nullable */
  lastReadingDate?: string | null;
  dataClassification: string;
  createdAt: Date;
}